requirements_dev.txt
# challenge dependencies
SimpleITK==2.5.0
numpy<2
timm==1.0.15
torchvision>=0.20

# custom dependencies
omegaconf==2.3.0
hydra-core==1.3.2
datasets==3.6.0
scikit-learn==1.7.0
matplotlib==3.10.3
torch>=2
lightning>=2
tensorboard==2.19.0

